requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.9.0
api==0.0.7
bcrypt==4.1.2
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
config==0.5.1
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==1.3.1
fastapi==0.99.1
fastapi-jwt-auth==0.5.0
fastapi-mail==1.3.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.1
nose==1.3.7
passlib==1.7.4
pip==25.0.1
pyasn1==0.6.1
pydantic==1.10.22
pydantic_core==2.33.2
pydantic-settings==2.10.1
PyJWT==1.7.1
pymongo==4.13.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
